pyproject.toml
[tool.poetry]
name = "sql-fake-populate"
version = "0.1.0"
description = ""
authors = ["Dyllon <dyllon.g@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pymssql = "^2.2.4"
Faker = "^13.3.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

